+++ console/commands/down.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+set -eu
+
+${DOCKER_COMPOSE} down "$@"