+++ console/commands/magento.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+set -eu
+
+${COMMANDS_DIR}/exec.sh php ${MAGENTO_DIR}/bin/magento "$@"