+++ console/commands/start.sh
@@ -1,0 +1,10 @@
+#!/usr/bin/env bash
+set -eu
+
+printf "${GREEN}Starting containers in detached mode${COLOR_RESET}\n"
+
+if [ "$#" == 0 ]; then
+    ${DOCKER_COMPOSE} up -d ${SERVICE_APP}
+else
+    ${DOCKER_COMPOSE} up -d "$@"
+fi