+++ console/tasks/list_functions.sh
@@ -1,0 +1,17 @@
+#!/usr/bin/env bash
+set -eu
+
+in_list()
+{
+    NEEDLE="$1"
+    LIST="$2"
+
+    # IMPORTANT: trailing spaces are needed to ensure matching condition:
+    # - https://stackoverflow.com/questions/8063228/how-do-i-check-if-a-variable-exists-in-a-list-in-bash
+    if [[ " ${LIST} " == *" ${NEEDLE} "* ]]; then
+        echo true
+        return
+    fi
+    echo false
+    return
+}